Added Controllers
Added
    RoleController
    BusController
    CityController
Added Expection handeling to Entites Not found situations.
Changed Long to Integer in Repositories

+++ src/main/java/com/bsuir/buspark/dal/BusRepository.java
@@ -4,5 +4,5 @@ import com.bsuir.buspark.entity.Bus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface BusRepository extends JpaRepository<Bus, Long> {
+public interface BusRepository extends JpaRepository<Bus, Integer> {
 }
